fix: replace invisible swipe zone with visible tap handle (#190)
The FilterEdgeSwipe component used an invisible swipe-up gesture zone
at the bottom of the viewport, which conflicted with iOS's native
home-indicator swipe gesture. Replace it with a visible tappable
bottom handle pill (BottomHandle) that uses tap instead of swipe,
eliminating the iOS gesture conflict entirely.

Closes #190
