--- packages/web/components/list/FilterEdgeSwipe.module.css
@@ -1,24 +1,0 @@
-.zone {
-  position: fixed;
-  left: 0;
-  /* Invisible gesture zone flush with the bottom of the viewport.
-     Sits just above the safe-area inset so swipe-up gestures are
-     captured before they reach Safari's toolbar. */
-  bottom: env(safe-area-inset-bottom, 0px);
-  z-index: 40;
-  width: 100%;
-  min-height: 20px;
-  padding: 0;
-  background: transparent;
-  border: none;
-  cursor: default;
-  /* Capture vertical pan gestures so upward swipes reach our handlers
-     rather than scrolling the list behind. */
-  touch-action: none;
-}
-
-@media (min-width: 768px) and (hover: hover) {
-  .zone {
-    display: none;
-  }
-}
